Fix various things
Fix printf commands to not expand the format string

Update the bin file to allow commands like docker-compose to still work.
The distinguishing factor there is that docker-compose require files
from the original CWD to work. As we used to cd away from the user's
CWD, then this command would not work.

A job for later is to get rid of the dependency of pipenv.

Change-Id: Ibf8a2ffd811e57ccc1ad88d1e0c78c6f68c43cd1